requirements.txt
# Core dependencies
google-adk>=0.1.0
google-generativeai>=0.4.1
litellm>=1.0.0
python-dotenv==1.0.0
requests>=2.31.0
supabase>=2.22.2